==> scripts/01A-compare_cellranger_simpleaf.R <==
out<-'outputs/01A-compare_cellranger_simpleaf'
dir.create(out)
library(Seurat)

#n of cells 
#simpleaf
Jorganoid<-readRDS('../../../akg/APOE_Jorganoid_project/outputs/simpleaf/APOE_Jorganoid.rds')

#cellranger

#n of UMIs/ cells and median gene/cells


#n of cells / gene
